Optional put idempotency.
Put operations in object store are usually idempotent. However, that guarantee does not exist since this crate calls APIs that it does not control.

Some clients might want to take the risk of ensuring all `put` operations are idempotent, and retried when unexpected errors happen, like SlateDB.

This change introduces a new flag in `PutOptions` to let callers mark the `put` operation as idempotent, triggering the retry logic that the underlying HTTP client implements.

I believe this is a useful compromise that prevents callers from implementing their own retry logic, which is already implemented in a much more robust manner in the HTTP libraries that end up calling the remote object store.