Make RetryError and RequestError public.
It's impossible to downcast an Error::Generic into a RetryError when the object store returns a generic error because of some kind of http connection error because RetryError and RequestError are not public.

This change makes those two types public. That way if the object store returns a Error::Generic, you can downcast it to a RetryError to try to find the cause of the issue.